Add SpannerClient benchmark.
