+++ grpc-gcp/src/benchmarkTest/java/SpannerBenchmark.java
@@ -24,8 +24,8 @@ public final class SpannerBenchmark {
   private SpannerBenchmark() {}
 
   private static int[] parseArgs(String[] args) {
-    // {number of rpcs, number of threads, payload bytes, isGrpcgcp, isSpannerClient}
-    int[] vars = new int[] {100, 1, 4096000, 1, 1};
+    // {number of rpcs, number of threads, payload bytes, isGrpcgcp, ClientType}
+    int[] vars = new int[] {100, 1, 1024, 1, 2};
     boolean usage = false;
     for (String arg : args) {
       if (!arg.startsWith("--")) {
@@ -48,8 +48,12 @@ public final class SpannerBenchmark {
         if (parts[1].equals("false")) {
           vars[3] = 0;
         }
-      } else if (key.equals("spanner_client")) {
-        if (parts[1].equals("false")) {
+      } else if (key.equals("client")) {
+        if (parts[1].equals("SpannerClient")) {
+          vars[4] = 2;
+        } else if (parts[1].equals("v1.SpannerClient")) {
+          vars[4] = 1;
+        } else if (parts[1].equals("v1.SpannerGrpc")) {
           vars[4] = 0;
         }
       } else {
@@ -91,9 +95,17 @@ public final class SpannerBenchmark {
                 + "GcpManagedChannel:%b\n"
                 + "Use Spanner Client API:%b",
             vars[0], vars[1], vars[2], vars[3] == 1, vars[4] == 1));
-    if (vars[4] == 1) {
+    if (vars[4] == 2) {
+      // Use the client: import com.google.cloud.spanner.DatabaseClient.
       SpannerClientTestCases testCases =
           new SpannerClientTestCases(vars[3] == 1, vars[2], vars[0], vars[1]);
+      testCases.prepareTestData();
+      testCases.testUpdateData();
+      testCases.testRead();
+    } else if (vars[4] == 1) {
+      // Use the client generated by gapic: com.google.cloud.spanner.v1.SpannerClient.
+      SpannerClientV1TestCases testCases =
+          new SpannerClientV1TestCases(vars[3] == 1, vars[2], vars[0], vars[1]);
       testCases.prepareTestData();
       testCases.testListSessions();
       testCases.testExecuteSql();
@@ -101,6 +113,7 @@ public final class SpannerBenchmark {
       testCases.testRead();
       testCases.testMaxConcurrentStream();
     } else {
+      // Use the naive implementation of SpannerGrpc stub.
       SpannerTestCases testCases = new SpannerTestCases(vars[3] == 1, vars[2], vars[0], vars[1]);
       testCases.prepareTestData();
       testCases.testListSessions();
